Re-scope init at block logic
